refactor: implement task detail page and refactor layout
Add a dedicated task detail page to improve modularity and clarity. Refactor the home page layout to support nested routes by introducing a children prop, enabling better code reusability.
