+++ apps/nextjs/src/app/(home)/layout.tsx
@@ -1,5 +1,6 @@
+import type { PropsWithChildren } from 'react';
 import { Page } from '@/components/pages/home/index';
 
-export default function HomePage() {
-  return <Page />;
+export default function HomePage({ children }: PropsWithChildren) {
+  return <Page>{children}</Page>;
 }
